requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
ollama
faiss-cpu
python-docx
fpdf